xmimsim: brute history is now double array
seems to fix an issue on hpc when the brute history would not be
recorded in the xmso file.

+++ include/xmi_main.h
@@ -18,7 +18,7 @@ struct xmi_main_options {
 
 
 
-int xmi_main_msim (xmi_inputFPtr inputFPtr, xmi_hdf5FPtr hdf5FPtr, int n_mpi_hosts, double **channels, int nchannels, struct xmi_main_options, long int **brute_history, double **var_red_history, struct xmi_solid_angle *solid_angles);
+int xmi_main_msim (xmi_inputFPtr inputFPtr, xmi_hdf5FPtr hdf5FPtr, int n_mpi_hosts, double **channels, int nchannels, struct xmi_main_options, double **brute_history, double **var_red_history, struct xmi_solid_angle *solid_angles);
 
 void xmi_detector_convolute(xmi_inputFPtr inputFPtr, xmi_hdf5FPtr hdf5FPtr, double *channels_noconv, double **channels_conv, int nchannels, struct xmi_main_options);
